Разделяет верстку на шаблоны
