+++ functions.php
@@ -58,7 +58,7 @@ function getAllTasks()
  * @param string $name имя категории
  * @return int количество сопадающих категорий
  */
-function counts_category_in_projects(array $projects_list, string $name): int
+function counts_category_in_projects(array $projects_list = [], string $name = ''): int
 {
     return array_reduce($projects_list, function ($carry, $item_project) use ($name) {
         $carry += $item_project['category'] === $name ? 1 : 0;
